feat: initialize websocket server
